% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.r
\name{as_dataframe_of_labels}
\alias{as_dataframe_of_labels}
\title{Convert a linked data frame to labels}
\usage{
as_dataframe_of_labels(d, ...)
}
\arguments{
\item{d}{A linked data frame}

\item{...}{Additional arguments passed to \code{data.frame}}
}
\value{
A data frame with labels in place of any ldf vectors
}
\description{
This take a data frame containing \code{ldf_resource} or \code{ldf_interval} vectors
and converts those vectors into labels.
}
\details{
The labels will either be character vectors or factors depending on the value of
\code{default.stringsAsFactors()}. This can be ridden by passing the argument
\code{stringsAsFactors=F}.
}
\examples{
uris <- c("http://example.net/id/apple",
          "http://example.net/id/banana",
          "http://example.net/id/carrot")
labels <- c("Apple","Banana","Carrot")
description <- data.frame(uri=uris, label=labels)

linked_data_frame <- data.frame(fruit=resource(uris, description))
labelled_data_frame <- as_dataframe_of_labels(linked_data_frame, stringsAsFactors=FALSE)
}
