% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.r
\name{as_dataframe_of_resources}
\alias{as_dataframe_of_resources}
\title{Enrich a data frame of URIs with resources descriptions}
\usage{
as_dataframe_of_resources(d, endpoint = default_endpoint())
}
\arguments{
\item{d}{A data frame containing URIs}

\item{endpoint}{A SPARQL endpoint}
}
\value{
A data frame of resources
}
\description{
This takes a data frame containing URIs and attempts to download descriptions for them.
}
\details{
The basic version just uses \code{\link[=get_label]{get_label()}}.
}
