% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.r
\name{curie}
\alias{curie}
\title{Extract Compact URI from resources}
\usage{
curie(resource, prefixes = default_prefixes())
}
\arguments{
\item{resource}{A vector of \code{ldf_resource}s}

\item{prefixes}{A named character vector mapping from prefix to namespace}
}
\value{
A character vector of compact URIs.
}
\description{
The URI is compacted using the specified prefixes.
If no prefixes are specified the defaults are taken from \code{default_prefixes} which can be
overriden by setting the global option "ldf_prefixes".
}
