% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmd.r
\name{default_endpoint}
\alias{default_endpoint}
\title{Retrieve the default sparql endpoint from configuration}
\usage{
default_endpoint()
}
\value{
returns the value of \code{ldf.default_endpoint} option,
defaulting to \verb{https://staging.gss-data.org.uk/sparql}
}
\description{
Retrieve the default sparql endpoint from configuration
}
\examples{
\dontrun{
default_endpoint() # returns default

options(ldf.default_endpoint="https://beta.gss-data.org.uk/sparql")
default_endpoint()
}
}
