% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.r
\name{description}
\alias{description}
\title{Extract the description from a resource}
\usage{
description(resource)
}
\arguments{
\item{resource}{A vector of \code{ldf_resource}s}
}
\value{
The resource's description - a type inheriting from data frame
}
\description{
Extract the description from a resource
}
