% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{download_progress}
\alias{download_progress}
\title{Execute SPARQL select query and parse results}
\usage{
download_progress(progress_env = new.env())
}
\arguments{
\item{progress_env}{an environment across which to retain the progress bar}
}
\value{
A progress bar function for use with httr
}
\description{
Execute SPARQL select query and parse results
}
