% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.r
\name{fill_missing}
\alias{fill_missing}
\title{Add rows to description for missing URIs}
\usage{
fill_missing(description, uri)
}
\arguments{
\item{description}{A resource description (inherits from data frame)}

\item{uri}{A character vector of URIs}
}
\value{
A resource description with all URIs included
}
\description{
The missing URIs will only be described with their URI
}
