% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmd.r
\name{get_codelist}
\alias{get_codelist}
\title{Download a Codelist}
\usage{
get_codelist(codelist_uri, endpoint = default_endpoint())
}
\arguments{
\item{codelist_uri}{A string}

\item{endpoint}{A string for the sparql endpoint}
}
\value{
A data frame
}
\description{
Returns a data frame with one observation per code and columns for the label, notation, and
sort priority (if available).
}
\examples{
\dontrun{
get_codelist("http://statistics.gov.scot/def/concept-scheme/ethnicity",
             "https://statistics.gov.scot/sparql")
}
}
