% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmd.r
\name{get_dimensions}
\alias{get_dimensions}
\title{Download dimension properties}
\usage{
get_dimensions(dataset_uri, endpoint = default_endpoint())
}
\arguments{
\item{dataset_uri}{A string}

\item{endpoint}{A string for the sparql endpoint}
}
\value{
A data frame containing the dimension uris, labels and optionally codelists
}
\description{
Download dimension properties
}
\examples{
\dontrun{
get_cube("http://gss-data.org.uk/data/gss_data/covid-19/ons-online-price-changes-for-high-demand-products#dataset")
}
}
