% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmd.r
\name{get_label}
\alias{get_label}
\title{Download a Resource's label}
\usage{
get_label(uri, endpoint = default_endpoint())
}
\arguments{
\item{uri}{A character vector of URIs.}

\item{endpoint}{A string for the sparql endpoint}
}
\value{
A data frame with column's for the \code{uri} and \code{label}
}
\description{
Provides for a very generic description, just requesting the \code{rdfs:label}.
}
\examples{
\dontrun{
get_label("http://purl.org/linked-data/cube#measureType")
}
}
