% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.r
\name{interval}
\alias{interval}
\alias{is_interval}
\title{\code{interval} vector}
\usage{
interval(uri)

is_interval(x)
}
\arguments{
\item{uri}{A character vector of URIs}

\item{x}{An interval}
}
\value{
An S3 vector of class \code{ldf_interval}.
}
\description{
This represents an interval from the UK Government
\href{http://reference.data.gov.uk/def/intervals}{intervals ontology}.
}
\details{
\code{is_interval} returns \code{TRUE} if \code{x} is an interval and \code{FALSE} otherwise.
}
\examples{
year <- interval("http://reference.data.gov.uk/id/year/2020")
is_interval(year)
month <- interval("http://reference.data.gov.uk/id/month/2020-07")
day <- interval("http://reference.data.gov.uk/id/day/2020-04-22")
p7d <- interval("http://reference.data.gov.uk/id/gregorian-interval/2020-04-27T00:00:00/P7D")
}
