% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.r
\name{label}
\alias{label}
\title{Extract resource labels}
\usage{
label(x)
}
\arguments{
\item{x}{A vector of \code{ldf_resource}s}
}
\value{
A character vector of labels
}
\description{
Extract resource labels
}
