% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmd.r
\name{list_cubes}
\alias{list_cubes}
\title{List all DataCubes}
\usage{
list_cubes(endpoint = default_endpoint())
}
\arguments{
\item{endpoint}{A string for the sparql endpoint (defaults to the value of \code{default_endpoint})}
}
\value{
A data frame
}
\description{
Returns a data frame of cubes.
}
\examples{
\dontrun{
# e.g. find datasets with "trade" in their label
subset(list_cubes(), grepl("trade", label, ignore.case=T))
}
}
