% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.r
\name{merge_description}
\alias{merge_description}
\title{Combine resource descriptions}
\usage{
merge_description(x, y)
}
\arguments{
\item{x}{A resource description (inherits from data frame)}

\item{y}{A resource description (inherits from data frame)}
}
\value{
A resource description (inherits from data frame)
}
\description{
Merges descriptions on URI, and attempts to merge columns from either description (filling with
\code{NA} where the column exists only in one description).
}
\examples{
a <- resource("a", data.frame(uri="a", label="A"))
description(a) <- merge_description(description(a), data.frame(uri="a", vowel=TRUE))
}
