% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmd.r
\name{predobj_binding}
\alias{predobj_binding}
\title{Create bindings for a labelled property}
\usage{
predobj_binding(properties)
}
\arguments{
\item{properties}{A list or data frame with vectors named \code{uri} and \code{label}}
}
\value{
a character vector of predicate-object bindings
}
\description{
Creates a query fragment for predicate-object pairs.
The predicate is taken from the property's \code{uri}.
The object is a variable binding made from the property's \code{label}.
}
