% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.r
\name{property}
\alias{property}
\title{Extract a property value from a description}
\usage{
property(resource, p)
}
\arguments{
\item{resource}{A vector of \code{ldf_resource}s}

\item{p}{A property (column name from the description)}
}
\value{
A vector with the values of the property that apply to each resource
}
\description{
Extract a property value from a description
}
