% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmd.r
\name{query}
\alias{query}
\title{Execute SPARQL select query and parse results}
\usage{
query(query_string, endpoint = default_endpoint(), format = "csv")
}
\arguments{
\item{query_string}{A string with the select query}

\item{endpoint}{A string defining the endpoint}

\item{format}{A string defining the format, either \code{"csv"} or \code{"json"}}
}
\value{
A data frame with a column for each variable binding, and a row per result
}
\description{
Execute SPARQL select query and parse results
}
\examples{
\dontrun{
query("SELECT * WHERE { ?s ?p ?o } LIMIT 10", "https://statistics.data.gov.uk/sparql")
}
}
