% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.r
\name{resource}
\alias{resource}
\alias{is_resource}
\title{\code{resource} vector}
\usage{
resource(
  uri = character(),
  description = data.frame(uri = unique(uri)),
  fill_missing = FALSE
)

is_resource(x)
}
\arguments{
\item{uri}{A character vector of URIs}

\item{description}{A data frame of descriptions (must have a \code{uri} column)}

\item{fill_missing}{A boolean specifying whether rows be added to the
description for missing URIs (defaults to \code{FALSE})}

\item{x}{Any vector}
}
\value{
An S3 vector of class \code{ldf_resource}.
}
\description{
A vector of RDF resources with descriptions held in an orthogonal table.
}
\examples{
resource("http://example.net")

uris <- c("http://example.net/id/apple",
          "http://example.net/id/banana",
          "http://example.net/id/carrot")
labels <- c("Apple","Banana","Carrot")
description <- data.frame(uri=uris, label=labels)
r <- resource(uris, description)
}
