% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.r
\name{sort_priority}
\alias{sort_priority}
\title{Extract resource sort priority}
\usage{
sort_priority(resource)
}
\arguments{
\item{resource}{A vector of \code{ldf_resource}s}
}
\value{
A vector of sort prorities (typically numeric)
}
\description{
Sort priority can be used to determine how to order resources
}
