% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.r
\name{uri}
\alias{uri}
\title{Extract the URI from a resource vector}
\usage{
uri(resource)
}
\arguments{
\item{resource}{A vector of \code{ldf_resource}s}
}
\description{
Extract the URI from a resource vector
}
