% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.r
\name{validate_resource}
\alias{validate_resource}
\title{Validate a resource}
\usage{
validate_resource(resource)
}
\arguments{
\item{resource}{A vector of \code{ldf_resource}s}
}
\value{
The resource (fluent interface for chaining)
}
\description{
Checks whether a URI column is present in the data descriptions and that it includes a value
for every resource with no duplicates.
}
\details{
Violations \code{stop} execution and report the error message.
}
