% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmd.r
\name{get_cube}
\alias{get_cube}
\title{Download a DataCube}
\usage{
get_cube(dataset_uri)
}
\arguments{
\item{dataset_uri}{A string}
}
\value{
A data frame
}
\description{
Returns a data frame with one observation per row and a column per component.
Components may be dimensions, measures or attributes.
}
\details{
Where the column represents an RDF Resource, it will have the type \code{ldf_resource} vector.
}
