% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmd.r
\name{get_dimensions}
\alias{get_dimensions}
\title{Download dimension properties}
\usage{
get_dimensions(dataset_uri, endpoint = default_endpoint())
}
\arguments{
\item{dataset_uri}{A string}

\item{endpoint}{A string for the sparql endpoint}
}
\value{
A data frame containing the dimension uris, labels and optionally codelists and parents
}
\description{
Download dimension properties
}
\examples{
\dontrun{
# Long URI is split here for legibility
base <- "http://gss-data.org.uk/data/gss_data/"
cube_uri <- paste0(base,"covid-19/ons-online-price-changes-for-high-demand-products#dataset")

get_cube(cube_uri)
}
}
