% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmd.r
\name{get_geography}
\alias{get_geography}
\title{Download Statistical Geographies}
\usage{
get_geography(geography_uri)
}
\arguments{
\item{geography_uri}{A character vector of URIs}
}
\value{
A data frame of geography descriptions
}
\description{
Connects to the \href{http://statistics.data.gov.uk/sparql}{ONS Geography Linked Date SPARQL interface} to
retrieve descriptions statistical geographies.
}
\details{
The description includes the official name, GSS-code and optionally the URI of the parent geography and
WKT geometry (if it exists).
}
\examples{
\dontrun{
get_geography("http://statistics.data.gov.uk/id/statistical-geography/K02000001")
}
}
