% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.r
\name{interval}
\alias{interval}
\alias{is_interval}
\title{\code{interval} vector}
\usage{
interval(uri)

is_interval(x)
}
\arguments{
\item{uri}{A character vector of URIs}
}
\value{
An S3 vector of class \code{ldf_interval}.
}
\description{
This represents an interval from the
\href{http://reference.data.gov.uk/def/intervals}{reference.data.gov.uk intervals ontology}
}
\examples{
year <- interval("http://reference.data.gov.uk/id/year/2020")
month <- interval("http://reference.data.gov.uk/id/month/2020-07")
day <- interval("http://reference.data.gov.uk/id/day/2020-04-22")
p7d <- interval("http://reference.data.gov.uk/id/gregorian-interval/2020-04-27T00:00:00/P7D")
}
