% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.r
\name{resource}
\alias{resource}
\alias{is_resource}
\title{\code{resource} vector}
\usage{
resource(uri = character(), data = data.frame(uri = unique(uri)))

is_resource(x)
}
\arguments{
\item{uri}{A character vector of URIs}

\item{data}{A data frame of descriptions (must have a \code{uri} column)}

\item{x}{Any vector}
}
\value{
An S3 vector of class \code{ldf_resource}.
}
\description{
This represents an RDF resource with the data held in an orthogonal table.
}
\examples{
resource("http://example.net")

uris <- c("http://example.net/id/apple",
          "http://example.net/id/banana",
          "http://example.net/id/carrot")
labels <- c("Apple","Banana","Carrot")
data <- data.frame(uri=uris, label=labels)
r <- resource(uris, data)
}
